crm-5: refactoring enum and trait files and move then to their own place
